sakhiBackend/requirements.txt
# requirements.txt
Django==5.1.1
djangorestframework==3.15.2
mongoengine==0.29.1
django-cors-headers==4.4.0
django-allauth==65.2.0
gunicorn==23.0.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
pymongo==4.8.0
python-dotenv==1.0.1
